fixed so that reminders send reliably
