+++ src/main/java/health/care/booking/services/ScheduleService.java
@@ -24,15 +24,7 @@ public class ScheduleService {
     @Autowired
     UserRepository userRepository;
 
-    @Scheduled(cron = "0 00 7 * * ?") // "Sec, Min, Hour"
-    public void sendDailyReminders() {
-        List<Appointment> appointmentList = appointmentService.getScheduledAppointmentsThatAreToday();
-        for (Appointment appointment : appointmentList) {
-                   User user = userRepository.findById(appointment.getPatientId()).orElseThrow(() -> new RuntimeException("No user found with id: " + appointment.getPatientId()));
-                   mailService.sendEmailReminder(user.getMail(), appointment.getDateTime().toString(), appointment.getReason());
-        }
-        System.out.println(appointmentList.size() + " Big list");
-    }
+
 }
